Handle any error from libvirt operations
A review of the helper code which wraps libvirtmod's compiled
module which implements the libvirt wrapper for Python, which
we refer to and call as libvirt or python-libvirt, reveals that
it appears possible for AttributeError and KeyError to be raised
up to the intermediate library level. While everything else refers
to libvirtError, it still seems possible, albeit remote, that
the generated code and interactions may still raise exceptions
besides libvirtError which previously would result in unexpected
behavior.

What we sometimes see on the ipmitool side is a generic
"Error: Unable to establish IPMI v2 / RMCP+ session" error which
is generated in many cases, including when no response is received.

So, since libvirtError is based on an Exception class, changes
exception handling to just catch and log based upon Exception.

Change-Id: I8159f5d1de2acb0678e7c85306413fab6999e615